refactor: dpll test structure
