+++ src/test/scala/satify/update/solver/dpll/cnf/CNFSimplificationTest.scala
@@ -1,10 +1,10 @@
-package satify.update.solver.dpll
+package satify.update.solver.dpll.cnf
 
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
 import satify.model.cnf.Bool.{False, True}
-import satify.model.cnf.CNF.*
 import satify.model.cnf.CNF
+import satify.model.cnf.CNF.*
 import satify.model.dpll.Constraint
 import satify.update.solver.dpll.cnf.CNFSimplification.*
